Added player and cpu colision, ball reset after scoring and random ball speed at start and after each goal

+++ Allegro_Pong/Allegro_Pong/ClassesForGameStates.h
@@ -4,10 +4,10 @@
 class Ball {
 
 	public:
-		int Yspeed;
-		int Xspeed;
-		int Yposition;
-		int Xposition;
+		float Yspeed;
+		float Xspeed;
+		float Yposition;
+		float Xposition;
 };
 
 class PlayerPaddle {
